+++ client/capi/twincat_link/c_api.cpp
@@ -3,7 +3,7 @@
 // Created Date: 08/03/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 23/05/2021
+// Last Modified: 04/06/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -13,7 +13,7 @@
 #include "./twincat_link.h"
 #include "twincat_link.hpp"
 
-void AUTDTwinCATLink(void** out) {
-  auto* link = LinkCreate(autd::link::TwinCATLink::create());
+void AUTDLinkTwinCAT(void** out) {
+  auto* link = LinkCreate(autd::link::TwinCAT::create());
   *out = link;
 }