+++ client/capi/twincat_link/twincat_link.h
@@ -3,7 +3,7 @@
 // Created Date: 20/02/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 18/05/2021
+// Last Modified: 04/06/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -14,5 +14,5 @@
 #include "../base/header.h"
 
 extern "C" {
-EXPORT_AUTD void AUTDTwinCATLink(void** out);
+EXPORT_AUTD void AUTDLinkTwinCAT(void** out);
 }