+++ client/examples/soem.cpp
@@ -3,7 +3,7 @@
 // Created Date: 05/11/2020
 // Author: Shun Suzuki
 // -----
-// Last Modified: 01/06/2021
+// Last Modified: 04/06/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -17,7 +17,7 @@
 
 std::string get_adapter_name() {
   size_t i = 0;
-  const auto adapters = autd::link::SOEMLink::enumerate_adapters();
+  const auto adapters = autd::link::SOEM::enumerate_adapters();
   for (auto&& [desc, name] : adapters) std::cout << "[" << i++ << "]: " << desc << ", " << name << std::endl;
 
   std::cout << "Choose number: ";
@@ -39,7 +39,7 @@ int main() {
     // If you have already recognized the EtherCAT adapter name, you can write it directly like below.
     // auto ifname = "\\Device\\NPF_{B5B631C6-ED16-4780-9C4C-3941AE8120A6}";
     const auto ifname = get_adapter_name();
-    auto link = autd::link::SOEMLink::create(ifname, autd->geometry()->num_devices());
+    auto link = autd::link::SOEM::create(ifname, autd->geometry()->num_devices());
     if (auto res = autd->open(std::move(link)); res.is_err()) {
       std::cerr << res.unwrap_err() << std::endl;
       return ENXIO;