+++ client/include/twincat_link.hpp
@@ -3,7 +3,7 @@
 // Created Date: 10/05/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 01/06/2021
+// Last Modified: 04/06/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -20,19 +20,19 @@ namespace autd::link {
 /**
  * @brief Link using TwinCAT
  */
-class TwinCATLink : public core::Link {
+class TwinCAT : public core::Link {
  public:
   /**
-   * @brief Create TwinCATLink
+   * @brief Create TwinCAT
    */
   static core::LinkPtr create();
 
-  TwinCATLink() = default;
-  ~TwinCATLink() override = default;
-  TwinCATLink(const TwinCATLink& v) noexcept = delete;
-  TwinCATLink& operator=(const TwinCATLink& obj) = delete;
-  TwinCATLink(TwinCATLink&& obj) = delete;
-  TwinCATLink& operator=(TwinCATLink&& obj) = delete;
+  TwinCAT() = default;
+  ~TwinCAT() override = default;
+  TwinCAT(const TwinCAT& v) noexcept = delete;
+  TwinCAT& operator=(const TwinCAT& obj) = delete;
+  TwinCAT(TwinCAT&& obj) = delete;
+  TwinCAT& operator=(TwinCAT&& obj) = delete;
 
   Error open() override = 0;
   Error close() override = 0;